backend/requirements.txt
# ─── Web Framework ────────────────────────────
fastapi==0.110.0          # The framework itself
uvicorn[standard]==0.29.0 # ASGI server that runs FastAPI (like nodemon but for Python)
slowapi==0.1.9            # Rate limiting (like express-rate-limit)

# ─── Database ─────────────────────────────────
sqlalchemy==2.0.29        # ORM (like Prisma, but Python)
alembic==1.13.1           # Database migrations (like prisma migrate)
asyncpg==0.29.0           # Async PostgreSQL driver

# ─── Auth & Security ──────────────────────────
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4            # Password hashing
python-multipart==0.0.9           # For form data / file uploads

# ─── Validation ───────────────────────────────
pydantic==2.6.4           # Data validation (FastAPI uses this automatically)
pydantic-settings==2.2.1  # Load .env into Python settings

# ─── Email ────────────────────────────────────
fastapi-mail==1.4.1       # Send emails easily

# ─── Payments ─────────────────────────────────
stripe==8.8.0

# ─── Image Uploads ────────────────────────────
cloudinary==1.39.1
python-magic==0.4.27      # Detect real file type (security)

# ─── Real-time (WebSockets) ───────────────────
websockets==12.0          # Built into FastAPI, this adds extra support

# ─── Utilities ────────────────────────────────
python-dotenv==1.0.1      # Load .env file
httpx==0.27.0             # Async HTTP client (for testing + external APIs)
redis==5.0.3              # Cache + pub/sub for real-time at scale

# ─── Dev Tools ────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6